apps/agents/pyproject.toml
[project]
name = "stepsquad-agents"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "google-cloud-bigquery>=3.25.0",
  "google-cloud-firestore>=2.16.0",
  "google-adk>=0.2.0",
  "google-generativeai>=0.8.0",
  "requests>=2.32.3",
  "python-dotenv>=1.0.0"
]
